refactor: extract context-menu logic into reusable mixin
